// Frontend/Módulo -Consulta de Simcards/control-archivo-simcard/src/app/models/SimcardPadre.ts
export interface SimcardPadre {
    id: number;
    fileName: string;
    orderDate: string;
    customerName: string;
    quantity: string;
    type: string;
    artWork: string;
    profile: string;
    batch: string;
    transportKey: string;
    simReference: string;
    graphRef: string;
    imsi: string;
    serNb: string;
    suppliersId: number;
    supplierName: string;
}